# apps/api/src/modules/beauty/beauty.controller.ts
import { Controller, Get, Put, Post, Param, Body, Query, Inject } from "@nestjs/common";
import { ApiTags, ApiBearerAuth, ApiBody, ApiParam, ApiQuery } from "@nestjs/swagger";
import { Roles, Session } from "@thallesp/nestjs-better-auth";
import { BeautyService } from "./beauty.service";
import { UpsertBeautyProfileDto, CreateShadeDto } from "../../dtos/beauty.dto";
import type { UserSession } from "../../common/types/session";

@ApiTags("Beauty Profiles")
@ApiBearerAuth()
@Controller("customers/:customerId")
export class BeautyController {
  constructor(@Inject(BeautyService) private beautyService: BeautyService) {}

  @Get("beauty-profile")
  @ApiParam({ name: "customerId", type: String })
  findProfile(@Param("customerId") customerId: string) {
    return this.beautyService.findProfile(customerId);
  }

  @Put("beauty-profile")
  @ApiParam({ name: "customerId", type: String })
  @ApiBody({ type: UpsertBeautyProfileDto })
  upsertProfile(
    @Param("customerId") customerId: string,
    @Body() body: UpsertBeautyProfileDto,
    @Session() session: UserSession,
  ) {
    return this.beautyService.upsertProfile(
      { ...body, customerId },
      session.user,
    );
  }

  @Post("shades")
  @ApiParam({ name: "customerId", type: String })
  @ApiBody({ type: CreateShadeDto })
  addShade(
    @Body() body: CreateShadeDto,
    @Session() session: UserSession,
  ) {
    return this.beautyService.addShade(body, session.user);
  }

  @Get("shade-matches")
  @ApiParam({ name: "customerId", type: String })
  @ApiQuery({ name: "category", type: String })
  @ApiQuery({ name: "brandId", type: String, required: false })
  getShadeMatches(
    @Param("customerId") customerId: string,
    @Query("category") category: string,
    @Query("brandId") brandId: string | undefined,
    @Session() session: UserSession,
  ) {
    return this.beautyService.getShadeMatches(
      customerId,
      category,
      brandId,
      session.user,
    );
  }
}