# apps/ios/LorealClienteling/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/brands": {
            "get": {
                "operationId": "BrandsController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Brands"
                ]
            },
            "post": {
                "operationId": "BrandsController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBrandDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Brands"
                ]
            }
        },
        "/brands/{id}": {
            "get": {
                "operationId": "BrandsController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Brands"
                ]
            },
            "patch": {
                "operationId": "BrandsController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBrandDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Brands"
                ]
            }
        },
        "/brands/{id}/config": {
            "put": {
                "operationId": "BrandsController_upsertConfig",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Brands"
                ]
            }
        },
...
        "/customers/{id}": {
            "get": {
                "operationId": "CustomersController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Customers"
                ]
            },
            "patch": {
                "operationId": "CustomersController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCustomerDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Customers"
                ]
            }
        },
...
        "/customers/{customerId}/beauty-profile": {
            "get": {
                "operationId": "BeautyController_findProfile",
                "parameters": [
                    {
                        "name": "customerId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Beauty Profiles"
                ]
            },
            "put": {
                "operationId": "BeautyController_upsertProfile",
                "parameters": [
                    {
                        "name": "customerId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpsertBeautyProfileDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Beauty Profiles"
                ]
            }
        },
        "/customers/{customerId}/shades": {
            "post": {
                "operationId": "BeautyController_addShade",
                "parameters": [
                    {
                        "name": "customerId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateShadeDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Beauty Profiles"
                ]
            }
        },
        "/customers/{customerId}/shade-matches": {
            "get": {
                "operationId": "BeautyController_getShadeMatches",
                "parameters": [
                    {
                        "name": "brandId",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "customerId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Beauty Profiles"
                ]
            }
        },
...
        "/appointments": {
            "get": {
                "operationId": "AppointmentsController_findAll",
                "parameters": [
                    {
                        "name": "to",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Appointments"
                ]
            },
            "post": {
                "operationId": "AppointmentsController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateAppointmentDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Appointments"
                ]
            }
        },
        "/appointments/calendar": {
            "get": {
                "operationId": "AppointmentsController_getCalendar",
                "parameters": [
                    {
                        "name": "to",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Appointments"
                ]
            }
        },
        "/appointments/{id}": {
            "get": {
                "operationId": "AppointmentsController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Appointments"
                ]
            },
            "patch": {
                "operationId": "AppointmentsController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAppointmentDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Appointments"
                ]
            }
        },
...
        "/audit-logs": {
            "get": {
                "operationId": "AuditController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Audit"
                ]
            }
        },
        "/audit-logs/{id}": {
            "get": {
                "operationId": "AuditController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "Audit"
                ]
            }
        },
...
    },
    "info": {
        "title": "L'Or\u00e9al Clienteling API",
        "description": "API for L'Or\u00e9al beauty advisor clienteling platform",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [
        {
            "url": "http://localhost:3001",
            "description": "Development"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateBrandDto": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "LANCOME",
                        "minLength": 1,
                        "maxLength": 50
                    },
                    "displayName": {
                        "type": "string",
                        "example": "Lanc\u00f4me",
                        "minLength": 1,
                        "maxLength": 200
                    },
                    "tier": {
                        "type": "string",
                        "enum": [
                            "luxury",
                            "premium",
                            "mass"
                        ],
                        "example": "luxury"
                    }
                },
                "required": [
                    "code",
                    "displayName",
                    "tier"
                ]
            },
            "UpdateBrandDto": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "LANCOME",
                        "minLength": 1,
                        "maxLength": 50
                    },
                    "displayName": {
                        "type": "string",
                        "example": "Lanc\u00f4me",
                        "minLength": 1,
                        "maxLength": 200
                    },
                    "tier": {
                        "type": "string",
                        "enum": [
                            "luxury",
                            "premium",
                            "mass"
                        ],
                        "example": "luxury"
                    },
                    "active": {
                        "type": "boolean"
                    }
                }
            },
...
            "UpdateCustomerDto": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string",
                        "example": "Mar\u00eda",
                        "minLength": 1,
                        "maxLength": 100
                    },
                    "lastName": {
                        "type": "string",
                        "example": "L\u00f3pez",
                        "minLength": 1,
                        "maxLength": 100
                    },
                    "email": {
                        "type": "string",
                        "example": "maria@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "5551234567",
                        "minLength": 10,
                        "maxLength": 15
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "female",
                            "male",
                            "non_binary",
                            "prefer_not_say"
                        ],
                        "example": "female"
                    },
                    "birthDate": {
                        "type": "string",
                        "format": "date-time",
                        "example": "1990-05-15T00:00:00.000Z"
                    }
                }
            },
            "UpsertBeautyProfileDto": {
                "type": "object",
                "properties": {
                    "customerId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "skinType": {
                        "type": "string",
                        "enum": [
                            "dry",
                            "oily",
                            "combination",
                            "sensitive",
                            "normal"
                        ]
                    },
                    "skinTone": {
                        "type": "string",
                        "enum": [
                            "fair",
                            "light",
                            "medium",
                            "tan",
                            "deep"
                        ]
                    },
                    "skinSubtone": {
                        "type": "string",
                        "enum": [
                            "cool",
                            "neutral",
                            "warm"
                        ]
                    },
                    "skinConcerns": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "acne",
                                "aging",
                                "pigmentation",
                                "dryness",
                                "sensitivity",
                                "pores",
                                "dark_circles",
                                "redness"
                            ]
                        }
                    },
                    "preferredIngredients": {
                        "example": [
                            "retinol",
                            "niacinamide"
                        ],
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "avoidedIngredients": {
                        "example": [
                            "alcohol",
                            "parabens"
                        ],
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "fragrancePreferences": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "floral",
                                "woody",
                                "citrus",
                                "oriental",
                                "fresh",
                                "gourmand"
                            ]
                        }
                    },
                    "makeupPreferences": {
                        "type": "object"
                    },
                    "routineType": {
                        "type": "string",
                        "enum": [
                            "morning",
                            "night",
                            "both"
                        ]
                    },
                    "interests": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "skincare",
                                "makeup",
                                "fragrance"
                            ]
                        }
                    }
                },
                "required": [
                    "customerId"
                ]
            },
            "CreateShadeDto": {
                "type": "object",
                "properties": {
                    "beautyProfileId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "category": {
                        "type": "string",
                        "enum": [
                            "foundation",
                            "concealer",
                            "lipstick",
                            "blush"
                        ],
                        "example": "foundation"
                    },
                    "brandId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "productId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "shadeCode": {
                        "type": "string",
                        "example": "N4.5",
                        "minLength": 1,
                        "maxLength": 50
                    }
                },
                "required": [
                    "beautyProfileId",
                    "category",
                    "brandId",
                    "productId",
                    "shadeCode"
                ]
            },
...
            "CreateAppointmentDto": {
                "type": "object",
                "properties": {
                    "customerId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "eventType": {
                        "type": "string",
                        "enum": [
                            "cabin_service",
                            "facial",
                            "anniversary_event",
                            "vip_cabin",
                            "product_followup",
                            "custom"
                        ],
                        "example": "cabin_service"
                    },
                    "scheduledAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "durationMinutes": {
                        "type": "number",
                        "example": 60,
                        "minimum": 1,
                        "maximum": 480
                    },
                    "comments": {
                        "type": "string",
                        "maxLength": 1000
                    },
                    "isVirtual": {
                        "type": "boolean",
                        "default": false
                    },
                    "videoLink": {
                        "type": "string",
                        "example": "https://meet.google.com/abc-defg-hij"
                    }
                },
                "required": [
                    "customerId",
                    "eventType",
                    "scheduledAt",
                    "durationMinutes"
                ]
            },
            "UpdateAppointmentDto": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "scheduled",
                            "confirmed",
                            "rescheduled",
                            "cancelled",
                            "completed",
                            "no_show"
                        ]
                    },
                    "scheduledAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "durationMinutes": {
                        "type": "number",
                        "minimum": 1,
                        "maximum": 480
                    },
                    "comments": {
                        "type": "string",
                        "maxLength": 1000
                    }
                }
            },
...
        }
    }
}